pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "openreview_downloader"
version = "0.1.4"
description = "Download OpenReview papers by decision."
readme = "README.md"
authors = [
  { name = "Miroslav Lzicar", email = "lzicar2000@gmail.com" },
]
license = { text = "Apache-2.0" }
keywords = ["neurips", "iclr", "icml", "openreview", "papers", "automation", "downloader", "scraping"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.11"
dependencies = [
  "openreview-py>=1.54.2,<2.0.0",
  "tqdm>=4.67.0,<5.0.0",
  "PySide6>=6.6.0",
  "SQLAlchemy>=2.0.0",
  "alembic>=1.13.0",
  "pydantic>=2.5.0",
  "PyYAML>=6.0.0",
  "httpx>=0.25.0",
  "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
  "bump2version",
  "ruff",
  "mypy",
  "pylint",
  "autopep8",
  "semgrep",
  "radon",
  "refurb",
  "pytest",
  "pytest-cov",
  "docformatter",
  "cohesion",
  "instructor",
  "pandera",
  "wily",
  "hypothesis",
  "mutmut",
  "py-spy",
  "sqlfluff",
  "jinja2",
  "deptry",
  "interrogate",
  "pydeps",
  "rope",
  "bandit",
  "xenon",
  "grimp",
  "pip-audit",
  "types-PyYAML",
  "types-tqdm",
]

[project.urls]
"Homepage" = "https://github.com/mireklzicar/openreview_downloader"
"Repository" = "https://github.com/mireklzicar/openreview_downloader"
"Issues" = "https://github.com/mireklzicar/openreview_downloader/issues"

[project.scripts]
ordl = "openreview_downloader.cli:main"
ordl-ui = "openreview_downloader.ui:main"

[tool.hatch.build.targets.wheel]
packages = ["src/openreview_downloader"]

[tool.hatch.build.targets.sdist]
include = [
  "openreview_downloader",
  "LICENSE",
  "README.md",
  "main.py",
]

[dependency-groups]
dev = [
    "semgrep>=1.0.0",
]
